refactor: Move WeaponAttackAdapter to VR module.
